# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "smirk-backend-core",
    "description": "Open, self-hostable backend for the Smirk non-custodial multi-chain wallet.",
    "contact": {
      "name": "John Winter Murphy",
      "email": "jw@such.software"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.3.0"
  },
  "paths": {
...
    "/wallet/grin/broadcast": {
      "post": {
        "tags": [
          "grin"
        ],
        "summary": "Broadcast a finalized, signed Grin transaction.",
        "operationId": "broadcast",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrinBroadcastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transaction broadcast",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrinBroadcastResponse"
                }
              }
            }
          },
          "400": {
            "description": "Grin disabled or malformed transaction"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/wallet/lws/balance": {
      "post": {
        "tags": [
          "xmr_wow"
        ],
        "summary": "Balance + scan state for an account.",
        "operationId": "balance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account balance and scan state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LwsBalanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset, address, or view key"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/lws/confirmations": {
      "post": {
        "tags": [
          "xmr_wow"
        ],
        "summary": "Confirmation count for a transaction (daemon).",
        "operationId": "confirmations",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Confirmation count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset or txid"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/lws/height": {
      "post": {
        "tags": [
          "xmr_wow"
        ],
        "summary": "Current chain height (daemon) for confirmation counting.",
        "operationId": "height",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LwsAssetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Chain height",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeightResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/lws/history": {
      "post": {
        "tags": [
          "xmr_wow"
        ],
        "summary": "Transaction history (confirmed + mempool) for an account.",
        "operationId": "history",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account transaction history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LwsHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset, address, or view key"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/lws/random_outs": {
      "post": {
        "tags": [
          "xmr_wow"
        ],
        "summary": "Random decoy outputs for ring construction.",
        "operationId": "random_outs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RandomOutsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Decoy outputs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RandomOutsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset or request"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/lws/register": {
      "post": {
        "tags": [
          "xmr_wow"
        ],
        "summary": "Register an account (its view key) with the LWS so it begins scanning.\nIdempotent at the LWS; pass `start_height` to scan from a wallet birthday.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account registered for scanning",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset, address, or view key"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/lws/submit_tx": {
      "post": {
        "tags": [
          "xmr_wow"
        ],
        "summary": "Broadcast a finalized, signed Monero/Wownero transaction.",
        "operationId": "submit_tx",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset or malformed tx hex"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/lws/unspent_outs": {
      "post": {
        "tags": [
          "xmr_wow"
        ],
        "summary": "Unspent outputs for an account (for client-side spend construction).",
        "operationId": "unspent_outs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Unspent outputs + fee parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspentOutsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset, address, or view key"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/utxo/balance": {
      "post": {
        "tags": [
          "btc_ltc"
        ],
        "summary": "Confirmed/unconfirmed balance for a BTC/LTC address.",
        "operationId": "balance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Address balance in satoshis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset or address"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/utxo/broadcast": {
      "post": {
        "tags": [
          "btc_ltc"
        ],
        "summary": "Broadcast a finalized, signed BTC/LTC transaction.",
        "operationId": "broadcast",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Broadcast accepted; returns the txid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset or malformed tx hex"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/wallet/utxo/history": {
      "post": {
        "tags": [
          "btc_ltc"
        ],
        "summary": "Confirmed + mempool transaction history for a BTC/LTC address.",
        "operationId": "history",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transaction history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset or address"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmationsRequest": {
        "type": "object",
        "description": "Confirmation-count query for a tx.",
        "required": [
          "asset",
          "txid"
        ],
        "properties": {
          "asset": {
            "type": "string"
          },
          "txid": {
            "type": "string",
            "description": "Transaction id (64 hex)."
          }
        }
      },
...
      "GrinBroadcastRequest": {
        "type": "object",
        "description": "Broadcast a finalized, signed Grin transaction (built + signed by the wallet).",
        "required": [
          "tx"
        ],
        "properties": {
          "tx": {
            "type": "object",
            "description": "The finalized transaction object (grin node `push_transaction` input)."
          }
        }
      },
      "GrinBroadcastResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "LwsAssetRequest": {
        "type": "object",
        "description": "Asset-only query.",
        "required": [
          "asset"
        ],
        "properties": {
          "asset": {
            "type": "string",
            "description": "`xmr` or `wow`."
          }
        }
      },
...
      "ViewRequest": {
        "type": "object",
        "description": "Address + private view key for a per-account query.",
        "required": [
          "asset",
          "address",
          "view_key"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "asset": {
            "type": "string",
            "description": "`xmr` or `wow`."
          },
          "view_key": {
            "type": "string",
            "description": "Private view key (64 hex). Forwarded to the LWS; never stored or logged."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session access token issued by the /auth/* endpoints."
      }
    }
  },
...
}